modal terminado
cambios globales

+++ backend/controllers/AplicacionesController.php
@@ -35,6 +35,8 @@ class AplicacionesController extends Controller
      */
     public function actionIndex()
     {
+      if(isset(Yii::$app->user->identity->id)){
+        if(SiteController::findCom(8) or SiteController::findCom(9) or SiteController::findCom(10) or SiteController::findCom(11)){
         $searchModel = new AplicacionesSearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
@@ -42,6 +44,13 @@ class AplicacionesController extends Controller
             'searchModel' => $searchModel,
             'dataProvider' => $dataProvider,
         ]);
+        }else {
+          $this->redirect(['site/error']);
+        }
+      }
+      else {
+        $this->redirect(['site/login']);
+      }
     }
 
     /**
@@ -52,9 +61,18 @@ class AplicacionesController extends Controller
      */
     public function actionView($id)
     {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
+      if(isset(Yii::$app->user->identity->id)){
+        if(SiteController::findCom(9)){
+          return $this->render('view', [
+              'model' => $this->findModel($id),
+          ]);
+        }
+        else {
+          $this->redirect(['site/error']);
+        }
+      }else {
+        $this->redirect(['site/login']);
+      }
     }
 
     /**
@@ -64,6 +82,8 @@ class AplicacionesController extends Controller
      */
     public function actionCreate()
     {
+      if(isset(Yii::$app->user->identity->id)){
+        if(SiteController::findCom(8)){
         $model = new Aplicaciones();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -73,8 +93,14 @@ class AplicacionesController extends Controller
         return $this->render('create', [
             'model' => $model,
         ]);
+      }
+      else {
+        $this->redirect(['site/error']);
+      }
+    }else {
+      $this->redirect(['site/login']);
     }
-
+  }
     /**
      * Updates an existing Aplicaciones model.
      * If update is successful, the browser will be redirected to the 'view' page.
@@ -84,22 +110,9 @@ class AplicacionesController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
-
-        //Se agrega para pasar de String a Array
-        $model->TiposId_fk5 = explode(',',$model->TiposId_fk5);
-        $model->TiposId_fk6 = explode(',',$model->TiposId_fk6);
-        $model->TiposId_fk7 = explode(',',$model->TiposId_fk7);
-        $model->TiposId_fk8 = explode(',',$model->TiposId_fk8);
-        $model->TiposId_fk9 = explode(',',$model->TiposId_fk9);
-        $model->TiposId_fk10 = explode(',',$model->TiposId_fk10);
-        $model->TiposId_fk12 = explode(',',$model->TiposId_fk12);
-        if (!empty($model->TiposId_fk14)) {
-          $model->TiposId_fk14 = explode(',',$model->TiposId_fk14);
-        }        
-        $model->TiposId_fk16 = explode(',',$model->TiposId_fk16);
-        $model->TiposId_fk18 = explode(',',$model->TiposId_fk18);
-        $model->TiposId_fk21 = explode(',',$model->TiposId_fk21);
+      if(isset(Yii::$app->user->identity->id)){
+        if(SiteController::findCom(10)){
+          $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
             return $this->redirect(['view', 'id' => $model->AppId]);
@@ -108,8 +121,14 @@ class AplicacionesController extends Controller
         return $this->render('update', [
             'model' => $model,
         ]);
+      }
+      else {
+        $this->redirect(['site/error']);
+      }
+    }else {
+      $this->redirect(['site/login']);
     }
-
+  }
     /**
      * Deletes an existing Aplicaciones model.
      * If deletion is successful, the browser will be redirected to the 'index' page.
@@ -119,10 +138,19 @@ class AplicacionesController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+      if(isset(Yii::$app->user->identity->id)){
+        if(SiteController::findCom(11)){
+          $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+          return $this->redirect(['index']);
     }
+    else {
+      $this->redirect(['site/error']);
+    }
+  }else {
+    $this->redirect(['site/login']);
+  }
+}
 
     /**
      * Finds the Aplicaciones model based on its primary key value.
@@ -139,6 +167,4 @@ class AplicacionesController extends Controller
 
         throw new NotFoundHttpException('The requested page does not exist.');
     }
-
-  public function loadFunctions(){}
 }
